ci: add kurtosis devnet test (#55)
* chore(kurtosis): pin teku image

Needed because of a current regression with latest image when being used from optimism-package which uses an old version of ethereum-package (4.5.0).

* chore(just): add commands to be used by ci

Also commented some of the targets

* ci: add kurtosis-devnet ci

* chore: fix mise.toml dependencies

cast was working because it was using my locally installed cast. But it didn't work in CI because mise.toml was wrongly configured for it. Fixed it in this PR.

* chore: fix mise.toml

* fix(kurtosis): reenable observability

There's a bug right now where observability can't be disabled. See https://github.com/ethpandaops/optimism-package/issues/221

* chore: disable cast warnings in just _kurtosis_wait_for_first_l2_finalized_block

* ci(kurtosis): build native client to speed up ci

* ci(kurtosis): explicitly wait for first l2 finalized block

Add this as a ci step so we can see how long it takes.

* ci(kurtosis): add rust cache