chore: fix asterisc (no-std) build (#65)
* chore: fix+refactor justfile to allow running asterisc

Previously asterisc command was just not working, calling justfile targets that didn't exist. Fixed that.

Also refactored such that we have a single justfile run-client target which can run against cannor or asterisc depending on an argument.

* chore(cargo.toml): add release-client-lto profile

Needed to build client to run on asterisc

* docs: update README with new justfile commands

* ci: fix ci with new kurtosis run command

* chore: refactor client to remove std deps

Goal is to be able to build the client for the asterisc target, which is a no-std dependency, whereas client right now was depending on std features.

looks like a huge change set but its really just 3 things:
1. add default-features = false to every dependency is cargo.toml
2. move witgen client from client-bin crate to its own create (its only used by op-succinct as a lib)
3. move https://github.com/bxue-l2/eigenda-v2-struct-rust into this repo as its own crate, so we can manage its dependency set (removed std feature from one of the deps, didnt seem used..)

* chore: make eigenda-v2-struct crate no-std

* chore: make proof crate no-std

* just: remove --no-default-features from build-client-for-asterisc target

* chore(cargo): patch rust-kzg-bn254 crates

Needed until https://github.com/Layr-Labs/rust-kzg-bn254/pull/54 is merged and a new crates release is created

* chore: make client crate no-std

* style: just fmt-native-fix

* style: remove unused deps

* ci(rust): add cargo build (native+asterisc) workflows

* ci: fix build target

Also refactor to using matrix strategy

* chore: update rust-kzg-bn254 deps to 0.1.2

* chore(client): remove prelude dependency

Seemed not needed..?